Fixed Bug #7631
having an empty category in the cash shop would lead to the item list duplicating itself on every map change (client fault), added a warning when parsing the file with a empty category, and the creation of a dummy apple to that category (priced at 999).
Special Thanks to Tepoo for all the information.
http://hercules.ws/board/tracker/issue-7631-cashshop-itemslist-increases-by-every-warpmapchange/

Signed-off-by: shennetsind <ind@henn.et>