Shadows System Base
http://hercules.ws/board/topic/581-implement-the-shadows-system/

Special Thanks to Yommy for all the client data, Haru for making it possible to get it out and rosfus for the details on the feature.
Shadows Feature requires packetver 20120925 or newer.

----------
Also: Updated all packets related to the shadows system, improved memory/processing of inventory/cart/storage/viewequip packets

Signed-off-by: shennetsind <ind@henn.et>