Updated char server's response to offline map server
Previously attempting to log in while map server was unavailable or not yet ready would close the client, now the char select window shows up and attempting to select a character while the map server is unavailable will display a pop up "the map is not available", client will remain in char select window and won't close anymore.
Special Thanks to Haruna.

Signed-off-by: shennetsind <ind@henn.et>