Clarify some Job Class vs MapID confusion (1/2)
This commit ensures that `sd->status.class` and related variables only
contain Job Classes (i.e. the client-compatible values, where High
Novice 4001) and are never checked agains the bitmask-based MapID
values.

As a rule of thumb, from now on, when a variable is named `class`, it
is intended to contain a Job Class ID and not a MapID. The type of such
variable shall be a signed `int16` or `int`.

To ensure that related third party code is also verified when this
commit is merged, the variable `struct mmo_charstatus::class_ (i.e.
`sd->status.class_`) is renamed to `class`.

Some issues in related lines are also fixed, including:

- A wrong check in the char server would prevent the correct detection
  of babies in code related to the family exp sharing.

- Baby Arch Bishops would not be affected by Eucharistica.

- A wrong check would cause the `questinfo()` script command not to
  display its information for most classes (except 1-1 classes).

- Map IDs and Job Classes were mixed up in `itemdb_jobid2mapid()` and
  `itemdb_jobmask2mapid()` for 1-1 classes (causing currently no harm,
  since they just happen to coincide).

- The Baby Sura class would not cause parties to be marked as
  containing a monk type character (for SLS Team Up purposes).

- Baby Geneticists would bypass the cart check when trying to equip
  ammunitions.

- Baby Mechanics would bypass the Mado Gear check when trying to equip
  ammunitions.

- Transcendent Shadow Chasers would lose the Stalkers' ability to clone
  transcendent skills.

Signed-off-by: Haru <haru@dotalux.com>