Clarify some Job Class vs MapID confusion (2/2)
This commit ensures that `sd->job` and related variables only contain
MapIDs (i.e. the serverside optimized values, where High Novice is
MAPID_NOVICE | JOBL_UPPER) and are never checked against the
client-based Job Class values.

As a rule of thumb, from now on, when a variable is named `job` or
`jobid`, it is intended to contain a MapID and not a Job Class ID. The
type of such variable shall be an unsigned `uint16` or `uint32`.

To ensure that related third party code is also verified when this
commit is merged, the variable `struct map_session_data::class_ (i.e.
`sd->class_`) is renamed to `job`.

Some issues in related lines are also fixed, including:

- The atcommand `@mount` would not check properly that the requesting
  character is a 2nd class Swordsman type.

- `pc->addfame()` would silently accept invalid MapIDs, sending unknown
  values to the client (more on this in a subsequent commit).

- `pc->famerank()` would not use the passed job as a bitmask, causing
  the caller to have to mask it beforehand (more on this in a
  subsequent commit).

- The Soul Linker check in TK_JUMPKICK wasn't future-proof (no harm
  caused currently).

- Gunslingers would be able to be targeted by Spiritual Bestowment
  (`MO_KITRANSLATION`) and Spiritual Sphere Absorption
  (`MO_ABSORBSPIRITS`) due to a faulty check introduced when the
  Rebellion class was implemented (causing unintended interaction with
  Coins).

Signed-off-by: Haru <haru@dotalux.com>