Ensure that pc->addfame() increments the correct fame points
The function now takes the rank type as argument, rather than guessing
it from the character's class. If the wrong fame point type for the
current character is requested, the request is ignored.

This fixes some (unofficial) edge cases where refining a signed item is
refined by a taekwon or an alchemist, they could obtain rank points.

Signed-off-by: Haru <haru@dotalux.com>