Fix a race condition when saving the storage on logout
The storage data is no longer cached in the char server. This removal
doesn't cause any performance loss, since the cached data was refreshed
(loaded from the database) after every save operation. This commit
moves the load at the beginning of the save rather than at the end, and
discars the cached copy after the save is completed.

Related to #1754

Signed-off-by: Haru <haru@dotalux.com>