Add type constants for the getiteminfo()/setiteminfo() buildins
Replacements are as follows:

    0 => ITEMINFO_BUYPRICE
    1 => ITEMINFO_SELLPRICE
    2 => ITEMINFO_TYPE
    3 => ITEMINFO_MAXCHANCE
    4 => ITEMINFO_SEX
    5 => ITEMINFO_LOC
    6 => ITEMINFO_WEIGHT
    7 => ITEMINFO_ATK
    8 => ITEMINFO_DEF
    9 => ITEMINFO_RANGE
    10 => ITEMINFO_SLOTS
    11 (Subtype, for weapons and ammunitions) => ITEMINFO_SUBTYPE
    11 (ViewSprite, for other item types) => ITEMINFO_VIEWSPRITE (NOT AVAILABLE YET)
    12 => ITEMINFO_ELV
    13 => ITEMINFO_WLV
    14 => ITEMINFO_VIEWID
    15 => ITEMINFO_MATK (NOT AVAILABLE YET - this was documented but never implemented)

Calls to getiteminfo() and setiteminfo() have been replaced with the
newly introduced constants. Other constants (such as W_ weapon subtypes)
in related code have been replaced as well, to improve code readability.

This fixes an issue in the Eden Tutorial script "Tutorial Goal", where
ITEMINFO_ATK was accidentally used instead of ITEMINFO_WEIGHT.

Note: calls to getiteminfo or setiteminfo with numeric type arguments in
third party scripts must be replaced with the respective constants. The
use of numeric literals is no longer recommended, and those values may
change in the future without notice. See the getiteminfo documentation
for details.

Signed-off-by: Haru <haru@dotalux.com>