Fix Chaotic Blessings never healing the enemy
It has a chance to roll for enemy and heal the enemy,
While calculating the chance was correct,
it was getting the enemy with battle->check_target.
But unit->stop_attack has been called before in
unit->skilluse_id2, making battle->check_target return NULL at that
point, since it has no attack target anymore.
So it defaulted to heal the Homunculus itself, instead of the
enemy.