Fix Homunculi being partially vaporized.
Vanilmirth for example would get incompletely vaporized when using
HVAN_EXPLOSION, if the owner died due to it, and then it would be killed at the end of it.
This produced a bug which made it not ressurectable but callable, once it
was called it would be dead though and you'd be unable to control it,
requiring you to ressurect it after that.

This bug got introduced by 0eca88efc69becc591428fbb6b9fddd9237afd3d .