add new config item_rate_add_chain
this allows server owners to fix a bug where rates declared in item_chain.conf
are not adjusted via `item_rate_adddrop`.  The new config was made in order to
avoid unexpected behavior to server owners since the bug has existed for
a while.

also account for this config in @reloadbattleconf by re-reading the chains