Update mailmap with some more mappings and re-generate AUTHORS with the latest commits
There are still missing mailmap mappings, and there may be incorrect ones.
If you'd like yours to be changed or added, please drop me a line (email or
forums) or open a pull request in the repository. Otherwise, I'll still be
cleaning it up little by little over the course of the next releases.

Signed-off-by: Haru <haru@dotalux.com>