Make it possible to disable the official guild storage system for recent packetvers
If DISABLE_OFFICIAL_GUILD_STORAGE is defined, the official guild storage
system won't be enabled and the custom one will be left in place, as to
not to require additional skill points to access existing storages.

if the OFFICIAL_GUILD_STORAGE flag is enabled on the command line, it is
sanity checked against the current PACKETVER to avoid enabling an
unsupported configuration.

Signed-off-by: Haru <haru@dotalux.com>