Add support for conditional comments in the script engine
Conditional comments are similar to #ifdef directives in the C
preprocessor, allowing to exclude parts of a script depending on whether
a certain flag is enabled or not.

The feature is designed to offer backward compatibility to scripts that
use it, letting older engines seamlessly ignore the conditional part as
if it was a comment.

The structure of a conditional comment is the following:

// The following block is only parsed if FLAG is defined. If not
// defined, or when loaded by an older engine, the block is treated as a
// block comment.

/*@FLAG
.@x = 1;
//@*/

// The following block is only parsed if FLAG is not defined. If
// defined, or when loaded by an older engine, the block is treated as a
// block comment.

/*@!FLAG
.@x = 1;
//@*/

// The following block is only parsed if FLAG is defined. If not
// defined, the block is treated as a block comment. When loaded by an
older engine, the block is also parsed.

/*@FLAG*/
.@x = 1;
//@*/

// The following block is only parsed if FLAG is not defined. If
// defined, the block is treated as a block comment. When loaded by an
// older engine, the block is also parsed.

/*@!FLAG*/
.@x = 1;
//@*/

The following flags are currently always defined:

- TRUE
- HERCULES

The following flag is currently never defined and can be used to
temporarily disable blocks of code:

- FALSE

The following flags are currently conditionally defined, based on the
renewal/pre-renewal compile time toggle:

- RENEWAL
- PRERENEWAL

Plugins can define conditional flags at runtime by calling
script->declare_conditional_feature("FLAG_NAME"), allowing scripts to
behave differently when a plugin is loaded.

Signed-off-by: Haru <haru@dotalux.com>