Add a setting to disable loading of GM scripts
This provides an easy way for people that think those scripts are a
security hazard to disable them completely.

This is based on the conditional comment syntax checking for the
LOADGMSCRIPTS flag. Third party scripts can also put their
administrative bits under LOADGMSCRIPTS conditional blocks in order to
let this setting skip them if desired.

Signed-off-by: Haru <haru@dotalux.com>