Remove safesnprintf(), replaced by snprintf()
snprintf is part of the c99 standard, which is required by Hercules.
This makes the custom function redundant.

The two functions are almost equivalent, except for their return value.
Instead of returning -1 when the buffer is too short, snprintf() returns
the amount of characters that would have been written if the buffer was
unbounded, except the '\0' terminato.

Caution should be used when updating existing third party code that
makes use of the return value.

Signed-off-by: Haru <haru@dotalux.com>