Replace online_char_data::server with an enum
It's no longer necessary to encode the server ID a character is
connected to since there is only one mapserver, so it's no longer
necessary to encode the values in an int but an enum makes more sense.

Some code branches that are now more clearly unreachable have also been
removed.

Signed-off-by: Haru <haru@dotalux.com>