fix userconfig API emotes default loading
Not having data and a server error were both being handled as no data,
which would lead a query error to overwrite user configs with defaults.

We now send a proper error when it fails to load due to an application
error. This makes official clients inform user about the issue and not
override their configs.

Signed-off-by: Guilherme G. Menaldo <guilherme.menaldo@outlook.com>