Sort and resynchronize VS project files between versions
Note: This was done without the VS IDE and comes with no guarantee that
it'll work.

My personal thought is that these files shouldn't be in the repository
at all since the way they are is just unmaintainable unless one has
several versions of VS installed and the patience to replicate each
change to the project file in all the version. VS (and Xcode) project
files shouldn't be part of the repository at all and should be replaced
with definitions for a generator tool (i.e. CMake) that can generate
projects for different IDEs that people might want to use.

Signed-off-by: Haru <haru@dotalux.com>