Fix AutoSpell not always setting the new skill
SC_AUTOSPELL val1 should be the AutoSpell skill level, not
the selected skill learned level.

This was broken during the refactor and is now fixed.

Giving it the selected skill level would prevent restarting the buff
mid way if the previous buff had a higher level for the target skill.

e.g. a Lv10 Fire Bolt would not be replaced by a Lv5 Earth Spikes