Fixed skill_attack using dmg.flag instead of dmg.dmg_lv in ATK_BLOCK check
In the shadow_flag branch of the delayed damage block, the counter_additional_effect
call was guarded with dmg.flag > ATK_BLOCK instead of dmg.dmg_lv > ATK_BLOCK,
meaning the damage level was never actually being checked. This also corrects the
style of the surrounding conditions to match the codebase convention.

Signed-off-by: Lorenzo Buitizon <the.keikun@gmail.com>